SummaryTask4/test/ua/nure/sereda/SummaryTask4/command/LogoutCommandTest.java
package ua.nure.sereda.SummaryTask4.command;

import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.runners.MockitoJUnitRunner;
import ua.nure.sereda.SummaryTask4.web.Path;
import ua.nure.sereda.SummaryTask4.web.command.common.LogoutCommand;

import javax.servlet.http.HttpServletRequest;
import javax.servlet.http.HttpServletResponse;
import javax.servlet.http.HttpSession;

import static org.junit.Assert.assertEquals;
import static org.mockito.Mockito.when;

@RunWith(MockitoJUnitRunner.class)
public class LogoutCommandTest {

    @Mock
    private HttpSession session;

    @Mock
    private HttpServletRequest request;

    @Mock
    private HttpServletResponse response;

    @InjectMocks
    private LogoutCommand logoutCommand;

    @Test
    public void test_execute() {

        when(request.getSession()).thenReturn(session);

        String expected = Path.PAGE_INDEX;
        String actual = logoutCommand.execute(request, response);
        assertEquals(expected, actual);
    }
}

